courses/MicroServices/code-examples/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
aio-pika==9.4.0
pydantic==2.5.3
python-dotenv==1.0.0

# Circuit Breaker
pybreaker==0.5.1

# Resilience
tenacity==8.2.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2

# Data Management
eventsourcing==6.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiomysql==0.2.0

# Deployment
docker==7.0.0
pyyaml==6.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
